aisearch/requirements.txt
python-dotenv
azure-search-documents==11.6.0b4
azure-identity
openai